Refactor: `useTokenInfo` exports
