+++ src/utils/hooks/index.ts
@@ -17,6 +17,8 @@ import { getMainClasses } from '~utils/css';
 
 import { RootStateRecord } from '../../modules/state';
 
+export { useTokenInfo, TokenInfoProvider } from './useTokenInfo';
+
 interface DataFetcher<S> {
   select: S;
   fetch: (...fetchArgs: any[]) => Action<any>;
